microservices.backup.20251215_101218/avatar-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
httpx==0.28.1
pytest==8.2.2
pytest-asyncio==0.23.7
moviepy==1.0.3
Pillow==10.2.0
numpy==1.26.3
requests==2.31.0
